+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-04-12 21:52:06
-// Commit: fix(perf): prevent double tap duplicate action in perf
+// Updated at: 2026-04-14 18:32:36
+// Commit: feat(perf): add async perf stream provider
 
 class perfComponentState {
-  final String commitRef = '357';
-  final String lastMessage = 'fix(perf): prevent double tap duplicate action in perf';
-  final DateTime timestamp = DateTime.parse('2026-04-12T21:52:06+05:00');
+  final String commitRef = '374';
+  final String lastMessage = 'feat(perf): add async perf stream provider';
+  final DateTime timestamp = DateTime.parse('2026-04-14T18:32:36+05:00');
 }